requirements.txt
annotated-types==0.6.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
greenlet==3.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
openpyxl==3.1.2
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
python-decouple==3.8
rich==13.7.0
shellingham==1.5.4
SQLAlchemy==2.0.25
typer==0.9.0
typing_extensions==4.9.0
